// client/src/models/upgrades/tools/tool.types.ts
export enum ToolType {
  SMALL_SHOVE = 'small shove',
  SMALL_PICKAXE = 'small pickaxe',

  SHOVE = 'shove',
  AXE = 'axe',
  PICKAXE = 'pickaxe',
  RAKE = 'rake',

  BIG_SHOVE = 'big shove',

  GOLD_SACK = 'gold sack',
  TURQUOISE_SACK = 'turquoise sack',
  ZAPHIRE_SACK = 'zaphire sack',
  EMERALD_SACK = 'emerald sack',
  RUBI_SACK = 'rubi sack',
  CAULDRON = 'cauldron',
}